Refactor image builder class and add code for flask model
